last 30 days amount spent fetched

+++ app/src/main/java/com/sparshchadha/expensetracker/feature/expense/presentation/ui/screens/ExpenseSettingsScreen.kt
@@ -44,7 +44,7 @@ import kotlin.math.abs
 
 @Composable
 fun ExpenseSettingsScreen(
-    currentExpenseBudget: Int,
+    currentExpenseBudget: Double,
     onChangeBudget: (Int) -> Unit,
     onBackPress: () -> Unit,
 ) {
