+++ app/src/main/java/com/sparshchadha/expensetracker/feature/profile/ui/compose/Profile.kt
@@ -49,7 +49,7 @@ fun Profile(
     onLogout: () -> Unit,
     navigateToNotificationsScreen: () -> Unit,
     userName: String,
-    expenseBudget: Int,
+    expenseBudget: Double,
 ) {
 
     var showNameUpdateDialog by rememberSaveable {
